LayoutTests/http/tests/websocket/tests/hybi/inspector/resolveWebSocket-expected.txt

Tests for the Network.resolveWebSocket command.


== Running test suite: Network.resolveWebSocket
-- Running test case: Network.resolveWebSocket.validRequestIdentifier
PASS: Payload should have type "object".
PASS: Payload should have className "WebSocket".
PASS: Payload should have description "WebSocket".

-- Running test case: Network.resolveWebSocket.invalidRequestIdentifier
PASS: Should produce an error.
Error: Missing web socket for given requestId
